+++ lib/common_widgets/coordinates/gcw_coords/gcw_coords.dart
@@ -6,8 +6,8 @@ import 'package:gc_wizard/application/permissions/user_location.dart';
 import 'package:gc_wizard/application/theme/theme.dart';
 import 'package:gc_wizard/common_widgets/buttons/gcw_iconbutton.dart';
 import 'package:gc_wizard/common_widgets/clipboard/gcw_clipboard.dart';
-import 'package:gc_wizard/tools/coords/_common/formats/dec/widget/degrees_lat_textinputformatter.dart';
-import 'package:gc_wizard/tools/coords/_common/formats/dec/widget/degrees_lon_textinputformatter.dart';
+import 'package:gc_wizard/common_widgets/coordinates/gcw_coords/coord_format_inputs/degrees_lat_textinputformatter.dart';
+import 'package:gc_wizard/common_widgets/coordinates/gcw_coords/coord_format_inputs/degrees_lon_textinputformatter.dart';
 import 'package:gc_wizard/common_widgets/coordinates/gcw_coords/gcw_coords_formatselector.dart';
 import 'package:gc_wizard/common_widgets/coordinates/gcw_coords/gcw_coords_paste_button.dart';
 import 'package:gc_wizard/common_widgets/dividers/gcw_text_divider.dart';
